# server/docs/swagger.yaml
definitions:
...
  github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.ChatMessage:
    properties:
      content:
        type: string
      from:
        description: self 或 other
        type: string
      status:
        description: sent, delivered, read
        type: string
      time:
        type: string
      type:
        description: text, image, voice
        type: string
    type: object
...
  model.Game:
    properties:
      category:
        type: string
      categoryIds:
        items:
          type: integer
        type: array
      createdAt:
        type: string
      icon:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      pageStyle:
        description: 页面风格配置，存储JSON格式数据
        type: string
      platform:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  model.GameCategory:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      label:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.Message:
    properties:
      content:
        type: string
      conversationId:
        description: 会话ID，用于分组聊天记录
        type: string
      createdAt:
        type: string
      fromUserId:
        type: integer
      id:
        type: integer
      read:
        type: boolean
      status:
        description: sent, delivered, read
        type: string
      time:
        type: string
      toUserId:
        type: integer
      type:
        description: text, image, voice, system
        type: string
      updatedAt:
        type: string
    type: object
...
  request.SendMessageRequest:
    properties:
      content:
        type: string
      type:
        description: text, image, voice, system
        enum:
        - text
        - image
        - voice
        - system
        type: string
    required:
    - content
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.9.0
paths:
...
  /playmate/conversations:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取会话列表
      tags:
      - Message
  /playmate/conversations/{id}/read:
    put:
      consumes:
      - application/json
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 标记成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 通过会话ID标记会话为已读
      tags:
      - Message
  /playmate/conversations/archive/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 归档成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 归档会话
      tags:
      - Message
  /playmate/conversations/read/{userId}:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 标记成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 标记会话为已读
      tags:
      - Message
...
  /playmate/game-categories:
    get:
      consumes:
      - application/json
      parameters:
      - description: 状态
        in: query
        name: status
        type: string
      - description: 父分类ID
        in: query
        name: parentId
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取游戏分类列表成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.GameCategory'
                  type: array
                pagination:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取游戏分类列表
      tags:
      - GameCategory
  /playmate/game-categories/{category}/games:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分类ID
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 根据分类获取游戏列表成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Game'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 根据分类获取游戏列表
      tags:
      - GameCategory
  /playmate/games:
    get:
      consumes:
      - application/json
      parameters:
      - collectionFormat: csv
        description: 分类ID列表
        in: query
        items:
          type: integer
        name: categoryIds
        type: array
      - description: 状态
        in: query
        name: status
        type: string
      - description: 平台
        in: query
        name: platform
        type: string
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取游戏列表成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Game'
                  type: array
                pagination:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取游戏列表
      tags:
      - Game
...
  /playmate/messages:
    get:
      consumes:
      - application/json
      parameters:
      - description: 类型
        in: query
        name: type
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 发送者ID
        in: query
        name: senderId
        type: integer
      - description: 接收者ID
        in: query
        name: receiverId
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
                pagination:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取消息列表
      tags:
      - Message
  /playmate/messages/{id}/read:
    put:
      consumes:
      - application/json
      parameters:
      - description: 消息ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 标记成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 标记消息为已读
      tags:
      - Message
  /playmate/messages/chat/{userId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ChatMessage'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取聊天消息
      tags:
      - Message
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      - description: 消息内容
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 发送成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Message'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 发送消息
      tags:
      - Message
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...